+++ src/main/java/com/example/capstone/order/dto/CartRequestDTO.java
@@ -1,0 +1,9 @@
+package com.example.capstone.order.dto;
+
+
+public class CartRequestDTO {
+
+    private Long itemId;
+
+    private Integer quantity;
+}
