+++ src/main/java/com/example/capstone/page/PageService.java
@@ -20,16 +20,16 @@ public class PageService {
 
     private final AlarmRepository alarmRepository;
     private final CartRepository cartRepository;
-    private final ItemService itemService;
+    private final ItemService itemServiceImpl;
 
     // TODO: 서비스에 의존성을 가져서 리팩토링 해야할 수도
     public PageResponseDTO.Main getMainPage(Long memberId) {
         Integer numberUnreadAlarms = alarmRepository.countAllNotConfirmedAlarm(memberId);
         Integer numberItems = cartRepository.countAllItemInCart(memberId);
 
-        ItemResponseDTO.ItemList imminentItemList = itemService.getImminentItemList(PAGE, SIZE);
-        ItemResponseDTO.ItemList popularItemList = itemService.getPopularItemList(PAGE, SIZE);
-        ItemResponseDTO.ItemList subscribedItemList = itemService.getSubscribedItemList(memberId, PAGE, SIZE);
+        ItemResponseDTO.ItemList imminentItemList = itemServiceImpl.getImminentItemList(PAGE, SIZE);
+        ItemResponseDTO.ItemList popularItemList = itemServiceImpl.getPopularItemList(PAGE, SIZE);
+        ItemResponseDTO.ItemList subscribedItemList = itemServiceImpl.getSubscribedItemList(memberId, PAGE, SIZE);
 
         return PageResponseDTO.Main.builder()
                 .numberUnreadAlarms(numberUnreadAlarms)
